Ensure that no sudo is required after the first time through as-admin.sh.

+++ deploy/as-admin.sh
@@ -8,63 +8,66 @@ OPENTREE_HOST=$1
 
 APTGET="sudo apt-get -q --assume-yes --no-install-recommends"
 
-# ---------- UPDATE ----------
+function apt_get_install {
+    if [ ! -r .updated ]; then
+	$APTGET update
+	touch .updated
+    fi
+    $APTGET install $*
+}
 
-if [ ! -r .updated ]; then
-    $APTGET update
-    touch .updated
-fi
+# ---------- UPDATE ----------
 
 if [ `which dialog`x = x ]; then
     # I was hoping this would help with apache2's configure step, but it doesn't
-    $APTGET install dialog
+    apt_get_install dialog
 fi
 
 # ---------- RSYNC ----------
 if [ `which rsync`x = x ]; then
-    $APTGET install rsync
+    apt_get_install rsync
 fi
 
 # ---------- GCC (for some python packages) ----------
 if [ `which gcc`x = x ]; then
-    $APTGET install gcc
+    apt_get_install gcc
 fi
 
 
 # ---------- G++ (for NCL, the nexus, newick converter used by the curation tool's import) ----------
 if [ `which g++`x = x ]; then
-    $APTGET install g++
+    apt_get_install g++
 fi
 
 # ---------- make (for NCL, the nexus, newick converter used by the curation tool's import) ----------
 if [ `which make`x = x ]; then
-    $APTGET install make
+    apt_get_install make
 fi
 
 # ---------- autoconf and automake for NCL (curation dependency) ----------
 if [ `which autoconf`x = x ]; then
-    $APTGET install autotools-dev
+    apt_get_install autotools-dev
 fi
 # ---------- autoconf and automake for NCL (curation dependency) ----------
 if [ `which automake`x = x ]; then
-    $APTGET install automake
+    apt_get_install automake
 fi
 
 # ---------- PYTHON-DEV (for some python packages) ----------
 if [ ! -r /usr/include/*/Python.h ]; then
-    $APTGET install python-dev
+    apt_get_install python-dev
 fi
 # -----G++ (for NCL, the nexus, newick converter used by the curation tool's import) -----
 if [ `which g++`x = x ]; then
-    $APTGET install g++
+    apt_get_install g++
 fi
 # ---------- autoconf and automake for NCL (curation dependency) ----------
 if [ `which autoconf`x = x ]; then
-    $APTGET install autotools-dev
+    apt_get_install autotools-dev
 fi
 # ---------- autoconf and automake for NCL (curation dependency) ----------
 if [ `which automake`x = x ]; then
-    $APTGET install automake
+    apt_get_install automake
 fi
 
 
@@ -72,7 +75,7 @@ fi
 if [ ! -r /etc/init.d/apache2 ]; then
     echo Installing apache httpd
     # Prompts "do you want to continue?"
-    $APTGET install apache2
+    apt_get_install apache2
     echo Done
 fi
 
@@ -98,13 +101,13 @@ fi
 # unzip is needed for unpacking web2py.  Somebody broke the 'which' program -
 # you can't just check the status code any more.
 if [ `which unzip`x = x ]; then
-    $APTGET install unzip
+    apt_get_install unzip
 fi
 
 # ---------- PIP ----------
 # Get pip
 if [ `which pip`x = x ]; then
-    $APTGET install python-pip
+    apt_get_install python-pip
 fi
 
 # ---------- LIBCURL + PYCURL ---------- 
@@ -114,7 +117,7 @@ if false; then
     # used by oti indexing script (make sure we have SSL support)
     if [ `which curl`x = x ] || [ `curl-config --feature | grep SSL`x = x ]; then
     #    sudo apt-cache search libcurl-dev
-	$APTGET install libcurl4-openssl-dev
+	apt_get_install libcurl4-openssl-dev
 	# NOTE that we'll pip-install pycurl inside our venv (in index-doc-store.sh)
     fi
 fi
@@ -122,13 +125,13 @@ fi
 # ---------- GIT ----------
 # Get git (so we can clone the opentree repo)
 if [ `which git`x = x ]; then
-    $APTGET install git
+    apt_get_install git
 fi
 
 # ---------- WSGI ----------
 # Get wsgi (apache / web2py communication)
 if [ ! -r /etc/apache2/mods-enabled/wsgi.load ]; then
-    $APTGET install libapache2-mod-wsgi
+    apt_get_install libapache2-mod-wsgi
 fi
 
 # AWS has python 2.7.3 built in, no need to install it.
@@ -136,13 +139,13 @@ fi
 # ---------- PYTHON VIRTUALENV ----------
 # Get virtualenv
 if [ `which virtualenv`x = x ]; then
-    $APTGET install python-virtualenv
+    apt_get_install python-virtualenv
 fi
 
 # ---------- JAVA ----------
 if [ `which javac`x = x ]; then
-    $APTGET install openjdk-7-jre 
-    $APTGET install openjdk-7-jdk
+    apt_get_install openjdk-7-jre 
+    apt_get_install openjdk-7-jdk
 fi
 
 # Cf. file 'activate' - should be the same
@@ -155,18 +158,18 @@ fi
 
 # ---------- MAVEN 3 ----------
 if [ `which mvn`x = x ]; then
-    $APTGET install maven
+    apt_get_install maven
 fi
 
 # ---------- LSOF ----------
 # neo4j needs this
 if [ `which lsof`x = x ]; then
-    $APTGET install lsof
+    apt_get_install lsof
 fi
 
 # ---------- NTP ----------
 if [ ! -r /etc/ntp.conf ]; then
-    $APTGET install ntp
+    apt_get_install ntp
 fi
 
 
@@ -189,12 +192,15 @@ fi
 # the default 'vhost'.  The opentree config file gets put into
 # place later on in the setup sequence.
 
-sudo rm -f /etc/apache2/sites-enabled/000-default
-(cd /etc/apache2/sites-enabled; \
- sudo ln -sf ../sites-available/opentree ./000-opentree)
+if [ -r /etc/apache2/sites-enabled/000-default ]; then
+    sudo rm -f /etc/apache2/sites-enabled/000-default
+    (cd /etc/apache2/sites-enabled; \
+     sudo ln -sf ../sites-available/opentree ./000-opentree)
+fi
 
 # Enable the HTTPS site only if our SSL certs are found; else disable it
-if [ -r /etc/ssl/certs/opentree/STAR_opentreeoflife_org.crt ]; then
+if [ ! -r /etc/apache2/sites-enabled/001-opentree-ssl -a \
+     -r /etc/ssl/certs/opentree/STAR_opentreeoflife_org.crt ]; then
     (cd /etc/apache2/sites-enabled; \
      sudo ln -sf ../sites-available/opentree-ssl ./001-opentree-ssl)
 else
@@ -218,9 +224,13 @@ if [ ! -e ~opentree/.ssh ]; then
     sudo chown -R opentree:opentree ~opentree
 fi
 
-HOSTFILE=~opentree/hostname
-cat <<EOF | sudo bash
-    echo "$OPENTREE_HOST" >$HOSTFILE
-    chmod go+r $HOSTFILE
-    chown opentree $HOSTFILE
+# Ideally this one would be done every time, but we want to avoid unsatisfiable sudo prompt demands.
+# TBD: this code ought to be done by the 'opentree' user, not in this file.
+if [ ! -r ~opentree/hostname ]; then
+    HOSTFILE=~opentree/hostname
+    cat <<EOF | sudo bash
+	echo "$OPENTREE_HOST" >$HOSTFILE
+	chmod go+r $HOSTFILE
+	chown opentree $HOSTFILE
 EOF
+fi
